Move multi test to testing folder.
